hm-rest-api/src/requirements.txt
Flask==2.2.5
mysql-connector-python-rf==2.2.2